- Refactored int to Long. Option to be null and broader range of possible id's supported

+++ src/main/java/manolovisoromero/person_processor/storage/PersistAdapter.java
@@ -15,5 +15,5 @@ public interface PersistAdapter<T extends Entity>{
 
     Collection<T> saveAndGetAll(T entity);
 
-    Optional<T> getById(Integer id);
+    Optional<T> getById(Long id);
 }
